Added Statistics module (Controller, Service, Repository, DTO)

+++ src/main/java/org/example/ewastev0_1/services/Interface/StatisticsService.java
@@ -6,7 +6,7 @@ import org.example.ewastev0_1.dto.response.LeaderboardResponse;
 import org.example.ewastev0_1.dto.response.UserStatisticsResponse;
 
 public interface StatisticsService {
-    EnvironmentalImpactResponse calculateEnvironmentalImpact(Long userId);
-    UserStatisticsResponse getUserStatistics(Long userId);
+    EnvironmentalImpactResponse calculateEnvironmentalImpact(Integer userId);
+    UserStatisticsResponse getUserStatistics(Integer userId);
     LeaderboardResponse getLeaderboard();
 }
